Extracted csv conversion into a class
